qc: Ensure type is uint32_t
Some C++ compiler complains about signed being compared with unsigned.

+++ include/maxscale/query_classifier.h
@@ -401,7 +401,7 @@ const char* qc_op_to_string(qc_query_op_t op);
  */
 static inline bool qc_query_is_type(uint32_t typemask, qc_query_type_t type)
 {
-    return (typemask & type) == type;
+    return (typemask & (uint32_t)type) == (uint32_t)type;
 }
 
 /**
